fixed some tests and fixed some more legacy path errors
