+++ examples/basic_env.py
@@ -3,7 +3,7 @@ import gym
 if __name__ == '__main__':
     env = gym.make('openmodelica_microgrid_gym:ModelicaEnv-v1',
                    max_episode_steps=None,
-                   model_output=dict(lc1=['inductor1.i', 'inductor2.i', 'inductor3.i']),
+                   net='../net/net.yaml',
                    model_path='../omg_grid/omg_grid.Grids.Network.fmu')
 
     env.reset()
