+++ examples/basic_env_norm.py
@@ -3,7 +3,7 @@ import gym
 if __name__ == '__main__':
     env = gym.make('openmodelica_microgrid_gym:ModelicaEnv_test-v1',
                    is_normalized=True,
-                   net='net.yaml',
+                   net='../net/net.yaml',
                    model_path='../omg_grid/omg_grid.Grids.Network.fmu')
 
     env.reset()
