+++ examples/simple_agent.py
@@ -11,7 +11,7 @@ class RndAgent(Agent):
 
 if __name__ == '__main__':
     env = gym.make('openmodelica_microgrid_gym:ModelicaEnv-v1',
-                   net='net.yaml',
+                   net='../net/net.yaml',
                    model_path='../omg_grid/omg_grid.Grids.Network.fmu')
 
     agent = RndAgent()
