+++ openmodelica_microgrid_gym/net/__init__.py
@@ -1,1 +1,4 @@
+from .components import MasterInverter, SlaveInverter, MasterInverterCurrentSourcing
 from .net import Network
+
+__all__ = ['Network', 'MasterInverter', 'SlaveInverter', 'MasterInverterCurrentSourcing']
