--- openmodelica_microgrid_gym/net/net.yaml
@@ -1,35 +1,0 @@
-v_nom: 230
-#freq_nom: 50
-ts: .5e-4
-
-components:
-  inv1:
-    #i_nom: 20
-    #i_lim: 30
-    #v_DC: 1000
-    id: inverter1
-    cls: MasterInverter
-    in:
-      u: [i1p1, i1p2, i1p3]    # names of the inputs
-    out:
-      v: [lc1.capacitor1.v,lc1.capacitor2.v, lc1.capacitor3.v]
-      i: [lc1.inductor1.i, lc1.inductor2.i, lc1.inductor3.i]
-    # iref: [0,0,0]
-    # vref: [1,0,0]
-  inv2:
-    id: inverter2
-    cls: SlaveInverter
-    #pll:
-    #  kP: 10
-    #  kI: 200
-    in:
-      u: [i2p1, i2p2, i2p3]
-    out:
-      v: [lcl1.capacitor1.v,lcl1.capacitor2.v, lcl1.capacitor3.v]
-      i: [lcl1.inductor1.i, lcl1.inductor2.i, lcl1.inductor3.i]
-    i_ref: [15,0,0]
-  load:
-    id: rl1
-    cls: Load
-    out:
-      i: [.inductor1.i, .inductor2.i, .inductor3.i]
